Rename     ManaPHP/Rpc/Http/Application.php (from ManaPHP/Rpc/Application.php) Rename     ManaPHP/Rpc/Http/Client.php (from ManaPHP/Rpc/Client.php) Rename     ManaPHP/Rpc/Http/Client/Adapter/Http.php (from ManaPHP/Rpc/Client/Adapter/Http.php) Rename     ManaPHP/Rpc/Http/Client/Adapter/Ws.php (from ManaPHP/Rpc/Client/Adapter/Ws.php) Rename     ManaPHP/Rpc/Http/Client/Exception.php (from ManaPHP/Rpc/Client/Exception.php) Rename     ManaPHP/Rpc/Http/Client/ProtocolException.php (from ManaPHP/Rpc/Client/ProtocolException.php) Rename     ManaPHP/Rpc/Http/Client/Service.php (from ManaPHP/Rpc/Client/Service.php) Rename     ManaPHP/Rpc/Http/ClientInterface.php (from ManaPHP/Rpc/ClientInterface.php) Rename     ManaPHP/Rpc/Http/Controller.php (from ManaPHP/Rpc/Controller.php) Rename     ManaPHP/Rpc/Http/Factory.php (from ManaPHP/Rpc/Factory.php) Rename     ManaPHP/Rpc/Http/Server.php (from ManaPHP/Rpc/Server.php) Rename     ManaPHP/Rpc/Http/Server/Adapter/Fpm.php (from ManaPHP/Rpc/Server/Adapter/Fpm.php) Rename     ManaPHP/Rpc/Http/Server/Adapter/Php.php (from ManaPHP/Rpc/Server/Adapter/Php.php) Rename     ManaPHP/Rpc/Http/Server/Adapter/Swoole.php (from ManaPHP/Rpc/Server/Adapter/Swoole.php) Rename     ManaPHP/Rpc/Http/Server/HandlerInterface.php (from ManaPHP/Rpc/Server/HandlerInterface.php) Rename     ManaPHP/Rpc/Http/ServerInterface.php (from ManaPHP/Rpc/ServerInterface.php) Modified   ManaPHP/Rpc/Command.php Modified   app-rpc/app/Application.php Modified   app-rpc/app/Controllers/Controller.php Modified   app-rpc/app/Services/TimeService.php
