+++ Rpc/Command.php
@@ -55,7 +55,7 @@ class Command extends \ManaPHP\Cli\Command
 
 namespace App\Services;
 
-use ManaPHP\Rpc\Client\Service;
+use ManaPHP\Rpc\Http\Client\Service;
 
 class $serviceName extends Service
 {
